+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/di/MatrixComponent.kt
@@ -22,6 +22,7 @@ import com.squareup.moshi.Moshi
 import dagger.BindsInstance
 import dagger.Component
 import im.vector.matrix.android.api.Matrix
+import im.vector.matrix.android.api.MatrixConfiguration
 import im.vector.matrix.android.api.auth.AuthenticationService
 import im.vector.matrix.android.internal.SessionManager
 import im.vector.matrix.android.internal.auth.AuthModule
@@ -47,6 +48,8 @@ internal interface MatrixComponent {
 
     fun context(): Context
 
+    fun matrixConfiguration(): MatrixConfiguration
+
     fun resources(): Resources
 
     fun olmManager(): OlmManager
@@ -63,6 +66,7 @@ internal interface MatrixComponent {
 
     @Component.Factory
     interface Factory {
-        fun create(@BindsInstance context: Context): MatrixComponent
+        fun create(@BindsInstance context: Context,
+                   @BindsInstance matrixConfiguration: MatrixConfiguration): MatrixComponent
     }
 }