+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/room/state/DefaultStateService.kt
@@ -27,7 +27,6 @@ import im.vector.matrix.android.api.session.events.model.EventType
 import im.vector.matrix.android.api.session.room.state.StateService
 import im.vector.matrix.android.api.util.Optional
 import im.vector.matrix.android.api.util.toOptional
-import im.vector.matrix.android.internal.crypto.MXCRYPTO_ALGORITHM_MEGOLM
 import im.vector.matrix.android.internal.database.mapper.asDomain
 import im.vector.matrix.android.internal.database.model.CurrentStateEventEntity
 import im.vector.matrix.android.internal.database.query.getOrNull
@@ -35,7 +34,6 @@ import im.vector.matrix.android.internal.database.query.whereStateKey
 import im.vector.matrix.android.internal.task.TaskExecutor
 import im.vector.matrix.android.internal.task.configureWith
 import io.realm.Realm
-import java.security.InvalidParameterException
 
 internal class DefaultStateService @AssistedInject constructor(@Assisted private val roomId: String,
                                                                private val monarchy: Monarchy,
@@ -76,23 +74,5 @@ internal class DefaultStateService @AssistedInject constructor(@Assisted private
                     this.callback = callback
                 }
                 .executeBy(taskExecutor)
-    }
-
-    override fun enableEncryption(algorithm: String, callback: MatrixCallback<Unit>) {
-        if (algorithm != MXCRYPTO_ALGORITHM_MEGOLM) {
-            callback.onFailure(InvalidParameterException("Only MXCRYPTO_ALGORITHM_MEGOLM algorithm is supported"))
-        } else {
-            val params = SendStateTask.Params(roomId,
-                    EventType.STATE_ROOM_ENCRYPTION,
-                    mapOf(
-                            "algorithm" to algorithm
-                    ))
-
-            sendStateTask
-                    .configureWith(params) {
-                        this.callback = callback
-                    }
-                    .executeBy(taskExecutor)
-        }
     }
 }